+++ lib/libxview/notify/nint_r_fd.c
@@ -14,7 +14,9 @@ static char     sccsid[] = "@(#)nint_r_fd.c 20.12 93/06/28 Copyr 1985 Sun Micro"
  * Nint_r_fd.c - Implement the nint_remove_fd_func private interface.
  */
 
-#include <xview_private/ntfy.h>
+#include <xview_private/nint_r_fd_.h>
+#include <xview_private/ndet_fd_.h>
+#include <xview_private/nintremove_.h>
 #include <xview_private/ndet.h>
 #include <xview_private/nint.h>
 
@@ -28,6 +30,6 @@ nint_remove_fd_func(nclient, func, type, fd)
     /* Check arguments */
     if (ndet_check_fd(fd))
 	return (notify_errno);
-    return (nint_remove_func(nclient, func, type, (NTFY_DATA) fd,
+    return (nint_remove_func(nclient, func, type, (NTFY_DATA)(long)fd,
 			     NTFY_USE_DATA));
 }
